Place BINDIR at the front of PATH for tests
BINDIR needs to be at the front of PATH to make sure that the "unit
tests" that are running singularity actually run the version that we are
building and not some other thing that happens to be on the PATH at this
time.

This is only a workaround for tests failing to use the buildcfg
information to find the correct singularity binary.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>